@@ -18,8 +18,5 @@ namespace CreativeSignalRGBBridge;
 
 public interface ICreativeDevice
 {
-    public static abstract string DeviceSelector
-    {
-        get;
-    }
+    public static abstract string DeviceSelector { get; }
 }
