@@ -1,5 +1,5 @@
 ﻿// This is the Creative SignalRGB Bridge Plugin/Service.
-// Copyright © 2023-2024 Harrison Boyd
+// Copyright © 2023-2025 Harrison Boyd
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -25,28 +25,17 @@ using Microsoft.Extensions.Logging;
 // Exit if not running as a windows service
 if (Environment.UserInteractive && !Debugger.IsAttached) return;
 
-
 var builder = Host.CreateApplicationBuilder(args);
-builder.Services.AddWindowsService(options =>
-{
-    options.ServiceName = "Creative SignalRGB Bridge";
-});
+builder.Services.AddWindowsService(options => { options.ServiceName = "Creative SignalRGB Bridge"; });
+
 builder.Logging.ClearProviders();
 builder.Logging.SetMinimumLevel(LogLevel.Trace);
-
-builder.Logging.AddEventLog(config =>
-{
-    config.SourceName = "Creative SignalRGB Bridge";
-});
-
-
+builder.Logging.AddEventLog(config => { config.SourceName = "Creative SignalRGB Bridge"; });
 
 //builder.Services.AddSingleton(typeof(ILogger<CreativeSignalRGBBridgeService>), typeof(ILogger<CreativeSignalRGBBridgeService>));
 
 builder.Services.AddSingleton(typeof(DeviceManager<>), typeof(DeviceManager<>));
 builder.Services.AddHostedService<CreativeSignalRGBBridgeService>();
 
-
 var host = builder.Build();
-
 host.Run();
